Created generic settings view
