+++ windows/appSettings.html
@@ -5,7 +5,7 @@
   <title>General Settings</title>
   <link rel="stylesheet" href="../styles/lib/pure.css">
   <link rel="stylesheet" href="../styles/main.css">
-  <link rel="stylesheet" href="../styles/views/app-settings-view.css">
+  <link rel="stylesheet" href="../styles/views/settings-view.css">
 
   <script src="../app/lib/jquery-2.1.4.js"></script>
   <script src="../app/lib/underscore.js"></script>
